refactor(styles): consolidate color variables and standardize component styles

+++ src/app/views/users/user-list/user-list.component.ts
@@ -64,6 +64,7 @@ export class UserListComponent implements OnInit, OnDestroy {
   deleteUser(user: User) {
     this.translate.get("USERS.DELETE.LABEL").subscribe((translations) => {
       this.bsModalRef = this.modalService.show(ConfirmModalComponent, {
+        class: "modal-confirm-custom",
         initialState: {
           title: translations.title,
           message: translations.message,
@@ -73,16 +74,19 @@ export class UserListComponent implements OnInit, OnDestroy {
           },
         },
       });
-    });
-    this.bsModalRef.content.onClose.subscribe((result) => {
-      if (result) {
-        this.userService.deleteUser(user.id).subscribe(() => {
-          this.translate
-            .get("USERS.DELETE.TOAST.USER_DELETED_SUCCESSFULLY")
-            .subscribe((translatedMessage: string) => {
-              this.toastrService.success(translatedMessage);
+
+      if (this.bsModalRef && this.bsModalRef.content) {
+        this.bsModalRef.content.onClose.subscribe((result) => {
+          if (result) {
+            this.userService.deleteUser(user.id).subscribe(() => {
+              this.translate
+                .get("USERS.DELETE.TOAST.USER_DELETED_SUCCESSFULLY")
+                .subscribe((translatedMessage: string) => {
+                  this.toastrService.success(translatedMessage);
+                });
+              this.getUsers();
             });
-          this.getUsers();
+          }
         });
       }
     });
